Move 'is-hidden' to private folder
