+++ addon/-private/execution_context/acceptance.js
@@ -1,4 +1,5 @@
 import {
+  findElement,
   findElementWithAssert,
   simpleFindElementWithAssert
 } from '../../helpers';
@@ -42,6 +43,10 @@ AcceptanceExecutionContext.prototype = {
 
   assertElementExists(selector, options) {
     simpleFindElementWithAssert(this.pageObjectNode, selector, options);
+  },
+
+  find(selector, options) {
+    return findElement(this.pageObjectNode, selector, options);
   },
 
   findWithAssert(selector, options) {
